refactor: move rating logic under it's own folder
