+++ projects/client/src/lib/sections/summary/components/rating/constants.ts
