+++ projects/client/src/lib/sections/summary/components/rating/mapRatingToSimpleRating.spec.ts
